+++ torrent.cpp
@@ -8,12 +8,12 @@ lt::torrent_handle Torrent::get_handle() {
 	return handle;
 }
 
-Torrent::Torrent(const unsigned int id) {
+Torrent::Torrent(const unsigned long int id) {
 	this->id = id;
 	// TODO - confirm ID does not already exists
 }
 
-const unsigned int Torrent::get_id() {
+const unsigned long int Torrent::get_id() {
 	return id;
 }
